Move IDocumentTextExtractor/IOneDriveService DI registration from KnowledgeBaseModule to SharedRagModule
