+++ backend/src/Anela.Heblo.Application/ApplicationModule.cs
@@ -58,7 +58,7 @@ public static class ApplicationModule
     public static IServiceCollection AddApplicationServices(this IServiceCollection services, IConfiguration configuration, IHostEnvironment? environment = null)
     {
         // Register shared RAG infrastructure
-        services.AddSharedRagModule();
+        services.AddSharedRagModule(configuration);
 
         // Register MediatR
         services.AddMediatR(cfg => cfg.RegisterServicesFromAssembly(typeof(ApplicationModule).Assembly));
